refactor(server): view feature flag v2
