+++ packages/twenty-server/src/engine/metadata-modules/view/controllers/view.controller.ts
@@ -14,8 +14,6 @@ import {
 import { type APP_LOCALES } from 'twenty-shared/translations';
 import { isDefined } from 'twenty-shared/utils';
 
-import { FeatureFlagKey } from 'src/engine/core-modules/feature-flag/enums/feature-flag-key.enum';
-import { FeatureFlagService } from 'src/engine/core-modules/feature-flag/services/feature-flag.service';
 import { I18nService } from 'src/engine/core-modules/i18n/i18n.service';
 import { WorkspaceEntity } from 'src/engine/core-modules/workspace/workspace.entity';
 import { AuthUserWorkspaceId } from 'src/engine/decorators/auth/auth-user-workspace-id.decorator';
@@ -40,7 +38,6 @@ import {
   ViewExceptionMessageKey,
 } from 'src/engine/metadata-modules/view/exceptions/view.exception';
 import { ViewRestApiExceptionFilter } from 'src/engine/metadata-modules/view/filters/view-rest-api-exception.filter';
-import { ViewV2Service } from 'src/engine/metadata-modules/view/services/view-v2.service';
 import { ViewService } from 'src/engine/metadata-modules/view/services/view.service';
 
 @Controller('rest/metadata/views')
@@ -49,8 +46,6 @@ import { ViewService } from 'src/engine/metadata-modules/view/services/view.serv
 export class ViewController {
   constructor(
     private readonly viewService: ViewService,
-    private readonly viewV2Service: ViewV2Service,
-    private readonly featureFlagService: FeatureFlagService,
     private readonly flatEntityMapsCacheService: WorkspaceManyOrAllFlatEntityMapsCacheService,
     private readonly i18nService: I18nService,
   ) {}
@@ -114,25 +109,10 @@ export class ViewController {
     @AuthWorkspace() workspace: WorkspaceEntity,
     @RequestLocale() locale?: keyof typeof APP_LOCALES,
   ): Promise<ViewDTO> {
-    const isWorkspaceMigrationV2Enabled =
-      await this.featureFlagService.isFeatureEnabled(
-        FeatureFlagKey.IS_WORKSPACE_MIGRATION_V2_ENABLED,
-        workspace.id,
-      );
-
-    let view: ViewDTO;
-
-    if (isWorkspaceMigrationV2Enabled) {
-      view = await this.viewV2Service.createOne({
-        createViewInput: input,
-        workspaceId: workspace.id,
-      });
-    } else {
-      view = await this.viewService.create({
-        ...input,
-        workspaceId: workspace.id,
-      });
-    }
+    const view = await this.viewService.createOne({
+      createViewInput: input,
+      workspaceId: workspace.id,
+    });
 
     const processedViews = await this.processViewsWithTemplates(
       [view],
@@ -152,31 +132,14 @@ export class ViewController {
     @AuthWorkspace() workspace: WorkspaceEntity,
     @AuthUserWorkspaceId() userWorkspaceId: string | undefined,
   ): Promise<ViewDTO> {
-    const isWorkspaceMigrationV2Enabled =
-      await this.featureFlagService.isFeatureEnabled(
-        FeatureFlagKey.IS_WORKSPACE_MIGRATION_V2_ENABLED,
-        workspace.id,
-      );
-
-    let updatedView: ViewDTO;
-
-    if (isWorkspaceMigrationV2Enabled) {
-      updatedView = await this.viewV2Service.updateOne({
-        updateViewInput: {
-          ...input,
-          id,
-        },
-        workspaceId: workspace.id,
-        userWorkspaceId,
-      });
-    } else {
-      updatedView = await this.viewService.update(
+    const updatedView = await this.viewService.updateOne({
+      updateViewInput: {
+        ...input,
         id,
-        workspace.id,
-        input,
-        userWorkspaceId,
-      );
-    }
+      },
+      workspaceId: workspace.id,
+      userWorkspaceId,
+    });
 
     const processedViews = await this.processViewsWithTemplates(
       [updatedView],
@@ -193,22 +156,10 @@ export class ViewController {
     @Param('id') id: string,
     @AuthWorkspace() workspace: WorkspaceEntity,
   ): Promise<{ success: boolean }> {
-    const isWorkspaceMigrationV2Enabled =
-      await this.featureFlagService.isFeatureEnabled(
-        FeatureFlagKey.IS_WORKSPACE_MIGRATION_V2_ENABLED,
-        workspace.id,
-      );
-
-    let deletedView: ViewDTO | null;
-
-    if (isWorkspaceMigrationV2Enabled) {
-      deletedView = await this.viewV2Service.deleteOne({
-        deleteViewInput: { id },
-        workspaceId: workspace.id,
-      });
-    } else {
-      deletedView = await this.viewService.delete(id, workspace.id);
-    }
+    const deletedView = await this.viewService.deleteOne({
+      deleteViewInput: { id },
+      workspaceId: workspace.id,
+    });
 
     return { success: isDefined(deletedView) };
   }
