+++ packages/twenty-server/src/engine/metadata-modules/view/services/view.service.ts
@@ -1,32 +1,365 @@
 import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
 
-import { type APP_LOCALES, SOURCE_LOCALE } from 'twenty-shared/translations';
 import { isDefined } from 'twenty-shared/utils';
-import { IsNull, Repository } from 'typeorm';
 
+import { InjectRepository } from '@nestjs/typeorm';
+import { ApplicationService } from 'src/engine/core-modules/application/application.service';
 import { I18nService } from 'src/engine/core-modules/i18n/i18n.service';
 import { generateMessageId } from 'src/engine/core-modules/i18n/utils/generateMessageId';
-import { FIND_ALL_CORE_VIEWS_GRAPHQL_OPERATION } from 'src/engine/metadata-modules/view/constants/find-all-core-views-graphql-operation.constant';
+import { WorkspaceManyOrAllFlatEntityMapsCacheService } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.service';
+import { computeFlatEntityMapsFromTo } from 'src/engine/metadata-modules/flat-entity/utils/compute-flat-entity-maps-from-to.util';
+import { findFlatEntityByIdInFlatEntityMapsOrThrow } from 'src/engine/metadata-modules/flat-entity/utils/find-flat-entity-by-id-in-flat-entity-maps-or-throw.util';
+import { fromCreateViewInputToFlatViewToCreate } from 'src/engine/metadata-modules/flat-view/utils/from-create-view-input-to-flat-view-to-create.util';
+import { fromDeleteViewInputToFlatViewOrThrow } from 'src/engine/metadata-modules/flat-view/utils/from-delete-view-input-to-flat-view-or-throw.util';
+import { fromDestroyViewInputToFlatViewOrThrow } from 'src/engine/metadata-modules/flat-view/utils/from-destroy-view-input-to-flat-view-or-throw.util';
+import { fromUpdateViewInputToFlatViewToUpdateOrThrow } from 'src/engine/metadata-modules/flat-view/utils/from-update-view-input-to-flat-view-to-update-or-throw.util';
+import { CreateViewInput } from 'src/engine/metadata-modules/view/dtos/inputs/create-view.input';
+import { DeleteViewInput } from 'src/engine/metadata-modules/view/dtos/inputs/delete-view.input';
+import { DestroyViewInput } from 'src/engine/metadata-modules/view/dtos/inputs/destroy-view.input';
+import { UpdateViewInput } from 'src/engine/metadata-modules/view/dtos/inputs/update-view.input';
+import { ViewDTO } from 'src/engine/metadata-modules/view/dtos/view.dto';
 import { ViewEntity } from 'src/engine/metadata-modules/view/entities/view.entity';
 import { ViewVisibility } from 'src/engine/metadata-modules/view/enums/view-visibility.enum';
-import {
-  ViewException,
-  ViewExceptionCode,
-  ViewExceptionMessageKey,
-  generateViewExceptionMessage,
-  generateViewUserFriendlyExceptionMessage,
-} from 'src/engine/metadata-modules/view/exceptions/view.exception';
-import { WorkspaceCacheStorageService } from 'src/engine/workspace-cache-storage/workspace-cache-storage.service';
+import { WorkspaceMigrationBuilderExceptionV2 } from 'src/engine/workspace-manager/workspace-migration-v2/exceptions/workspace-migration-builder-exception-v2';
+import { WorkspaceMigrationValidateBuildAndRunService } from 'src/engine/workspace-manager/workspace-migration-v2/services/workspace-migration-validate-build-and-run-service';
+import { APP_LOCALES, SOURCE_LOCALE } from 'twenty-shared/translations';
+import { IsNull, Repository } from 'typeorm';
 
 @Injectable()
 export class ViewService {
   constructor(
     @InjectRepository(ViewEntity)
     private readonly viewRepository: Repository<ViewEntity>,
+    private readonly workspaceMigrationValidateBuildAndRunService: WorkspaceMigrationValidateBuildAndRunService,
+    private readonly flatEntityMapsCacheService: WorkspaceManyOrAllFlatEntityMapsCacheService,
+    private readonly applicationService: ApplicationService,
     private readonly i18nService: I18nService,
-    private readonly workspaceCacheStorageService: WorkspaceCacheStorageService,
   ) {}
+
+  async createOne({
+    createViewInput,
+    workspaceId,
+    createdByUserWorkspaceId,
+  }: {
+    createViewInput: CreateViewInput;
+    workspaceId: string;
+    createdByUserWorkspaceId?: string;
+  }): Promise<ViewDTO> {
+    const { workspaceCustomFlatApplication } =
+      await this.applicationService.findWorkspaceTwentyStandardAndCustomApplicationOrThrow(
+        {
+          workspaceId,
+        },
+      );
+
+    const {
+      flatObjectMetadataMaps,
+      flatViewMaps: existingFlatViewMaps,
+      flatFieldMetadataMaps: existingFlatFieldMetadataMaps,
+    } = await this.flatEntityMapsCacheService.getOrRecomputeManyOrAllFlatEntityMaps(
+      {
+        workspaceId,
+        flatMapsKeys: [
+          'flatObjectMetadataMaps',
+          'flatViewMaps',
+          'flatFieldMetadataMaps',
+        ],
+      },
+    );
+
+    const flatViewFromCreateInput = fromCreateViewInputToFlatViewToCreate({
+      createViewInput,
+      workspaceId,
+      createdByUserWorkspaceId,
+      workspaceCustomApplicationId: workspaceCustomFlatApplication.id,
+    });
+
+    const validateAndBuildResult =
+      await this.workspaceMigrationValidateBuildAndRunService.validateBuildAndRunWorkspaceMigration(
+        {
+          fromToAllFlatEntityMaps: {
+            flatViewMaps: computeFlatEntityMapsFromTo({
+              flatEntityMaps: existingFlatViewMaps,
+              flatEntityToCreate: [flatViewFromCreateInput],
+              flatEntityToDelete: [],
+              flatEntityToUpdate: [],
+            }),
+          },
+          dependencyAllFlatEntityMaps: {
+            flatObjectMetadataMaps: flatObjectMetadataMaps,
+            flatFieldMetadataMaps: existingFlatFieldMetadataMaps,
+          },
+          buildOptions: {
+            isSystemBuild: false,
+          },
+          workspaceId,
+        },
+      );
+
+    if (isDefined(validateAndBuildResult)) {
+      throw new WorkspaceMigrationBuilderExceptionV2(
+        validateAndBuildResult,
+        'Multiple validation errors occurred while creating view',
+      );
+    }
+
+    const { flatViewMaps: recomputedExistingFlatViewMaps } =
+      await this.flatEntityMapsCacheService.getOrRecomputeManyOrAllFlatEntityMaps(
+        {
+          workspaceId,
+          flatMapsKeys: ['flatViewMaps'],
+        },
+      );
+
+    return findFlatEntityByIdInFlatEntityMapsOrThrow({
+      flatEntityId: flatViewFromCreateInput.id,
+      flatEntityMaps: recomputedExistingFlatViewMaps,
+    });
+  }
+
+  async updateOne({
+    updateViewInput,
+    workspaceId,
+    userWorkspaceId,
+  }: {
+    updateViewInput: UpdateViewInput;
+    workspaceId: string;
+    userWorkspaceId?: string;
+  }): Promise<ViewDTO> {
+    const {
+      flatViewMaps: existingFlatViewMaps,
+      flatFieldMetadataMaps: existingFlatFieldMetadataMaps,
+    } =
+      await this.flatEntityMapsCacheService.getOrRecomputeManyOrAllFlatEntityMaps(
+        {
+          workspaceId,
+          flatMapsKeys: ['flatViewMaps', 'flatFieldMetadataMaps'],
+        },
+      );
+
+    const flatViewFromUpdateInput =
+      fromUpdateViewInputToFlatViewToUpdateOrThrow({
+        updateViewInput,
+        flatViewMaps: existingFlatViewMaps,
+      });
+
+    const existingFlatView = existingFlatViewMaps.byId[updateViewInput.id];
+
+    // If changing visibility from WORKSPACE to UNLISTED, ensure createdByUserWorkspaceId is set
+    // This prevents the view from disappearing for the user making the change
+    if (
+      isDefined(existingFlatView) &&
+      isDefined(updateViewInput.visibility) &&
+      updateViewInput.visibility === 'UNLISTED' &&
+      existingFlatView.visibility === 'WORKSPACE' &&
+      isDefined(userWorkspaceId)
+    ) {
+      // Re-allocate the view to the current user
+      flatViewFromUpdateInput.createdByUserWorkspaceId = userWorkspaceId;
+    }
+
+    const validateAndBuildResult =
+      await this.workspaceMigrationValidateBuildAndRunService.validateBuildAndRunWorkspaceMigration(
+        {
+          fromToAllFlatEntityMaps: {
+            flatViewMaps: computeFlatEntityMapsFromTo({
+              flatEntityMaps: existingFlatViewMaps,
+              flatEntityToCreate: [],
+              flatEntityToDelete: [],
+              flatEntityToUpdate: [flatViewFromUpdateInput],
+            }),
+          },
+          dependencyAllFlatEntityMaps: {
+            flatFieldMetadataMaps: existingFlatFieldMetadataMaps,
+          },
+          buildOptions: {
+            isSystemBuild: false,
+          },
+          workspaceId,
+        },
+      );
+
+    if (isDefined(validateAndBuildResult)) {
+      throw new WorkspaceMigrationBuilderExceptionV2(
+        validateAndBuildResult,
+        'Multiple validation errors occurred while updating view',
+      );
+    }
+
+    const { flatViewMaps: recomputedExistingFlatViewMaps } =
+      await this.flatEntityMapsCacheService.getOrRecomputeManyOrAllFlatEntityMaps(
+        {
+          workspaceId,
+          flatMapsKeys: ['flatViewMaps'],
+        },
+      );
+
+    return findFlatEntityByIdInFlatEntityMapsOrThrow({
+      flatEntityId: updateViewInput.id,
+      flatEntityMaps: recomputedExistingFlatViewMaps,
+    });
+  }
+
+  async deleteOne({
+    deleteViewInput,
+    workspaceId,
+  }: {
+    deleteViewInput: DeleteViewInput;
+    workspaceId: string;
+  }): Promise<ViewDTO> {
+    const {
+      flatViewMaps: existingFlatViewMaps,
+      flatFieldMetadataMaps: existingFlatFieldMetadataMaps,
+    } =
+      await this.flatEntityMapsCacheService.getOrRecomputeManyOrAllFlatEntityMaps(
+        {
+          workspaceId,
+          flatMapsKeys: ['flatViewMaps', 'flatFieldMetadataMaps'],
+        },
+      );
+
+    const optimisticallyUpdatedFlatViewWithDeletedAt =
+      fromDeleteViewInputToFlatViewOrThrow({
+        deleteViewInput,
+        flatViewMaps: existingFlatViewMaps,
+      });
+
+    const validateAndBuildResult =
+      await this.workspaceMigrationValidateBuildAndRunService.validateBuildAndRunWorkspaceMigration(
+        {
+          fromToAllFlatEntityMaps: {
+            flatViewMaps: computeFlatEntityMapsFromTo({
+              flatEntityMaps: existingFlatViewMaps,
+              flatEntityToCreate: [],
+              flatEntityToDelete: [],
+              flatEntityToUpdate: [optimisticallyUpdatedFlatViewWithDeletedAt],
+            }),
+          },
+          dependencyAllFlatEntityMaps: {
+            flatFieldMetadataMaps: existingFlatFieldMetadataMaps,
+          },
+          buildOptions: {
+            isSystemBuild: false,
+          },
+          workspaceId,
+        },
+      );
+
+    if (isDefined(validateAndBuildResult)) {
+      throw new WorkspaceMigrationBuilderExceptionV2(
+        validateAndBuildResult,
+        'Multiple validation errors occurred while deleting view',
+      );
+    }
+
+    const { flatViewMaps: recomputedExistingFlatViewMaps } =
+      await this.flatEntityMapsCacheService.getOrRecomputeManyOrAllFlatEntityMaps(
+        {
+          workspaceId,
+          flatMapsKeys: ['flatViewMaps'],
+        },
+      );
+
+    return findFlatEntityByIdInFlatEntityMapsOrThrow({
+      flatEntityId: deleteViewInput.id,
+      flatEntityMaps: recomputedExistingFlatViewMaps,
+    });
+  }
+
+  async destroyOne({
+    destroyViewInput,
+    workspaceId,
+  }: {
+    destroyViewInput: DestroyViewInput;
+    workspaceId: string;
+  }): Promise<ViewDTO> {
+    const {
+      flatViewMaps: existingFlatViewMaps,
+      flatFieldMetadataMaps: existingFlatFieldMetadataMaps,
+    } =
+      await this.flatEntityMapsCacheService.getOrRecomputeManyOrAllFlatEntityMaps(
+        {
+          workspaceId,
+          flatMapsKeys: ['flatViewMaps', 'flatFieldMetadataMaps'],
+        },
+      );
+
+    const flatViewFromDestroyInput = fromDestroyViewInputToFlatViewOrThrow({
+      destroyViewInput,
+      flatViewMaps: existingFlatViewMaps,
+    });
+
+    const validateAndBuildResult =
+      await this.workspaceMigrationValidateBuildAndRunService.validateBuildAndRunWorkspaceMigration(
+        {
+          fromToAllFlatEntityMaps: {
+            flatViewMaps: computeFlatEntityMapsFromTo({
+              flatEntityMaps: existingFlatViewMaps,
+              flatEntityToCreate: [],
+              flatEntityToDelete: [flatViewFromDestroyInput],
+              flatEntityToUpdate: [],
+            }),
+          },
+          buildOptions: {
+            isSystemBuild: false,
+            inferDeletionFromMissingEntities: {
+              view: true,
+            },
+          },
+          dependencyAllFlatEntityMaps: {
+            flatFieldMetadataMaps: existingFlatFieldMetadataMaps,
+          },
+          workspaceId,
+        },
+      );
+
+    if (isDefined(validateAndBuildResult)) {
+      throw new WorkspaceMigrationBuilderExceptionV2(
+        validateAndBuildResult,
+        'Multiple validation errors occurred while destroying view',
+      );
+    }
+
+    return flatViewFromDestroyInput;
+  }
+
+  processViewNameWithTemplate(
+    viewName: string,
+    isCustom: boolean,
+    objectLabelPlural?: string,
+    locale?: keyof typeof APP_LOCALES,
+  ): string {
+    if (viewName.includes('{objectLabelPlural}') && objectLabelPlural) {
+      const messageId = generateMessageId(viewName);
+      const translatedTemplate = this.i18nService.translateMessage({
+        messageId,
+        values: {
+          objectLabelPlural,
+        },
+        locale: locale ?? SOURCE_LOCALE,
+      });
+
+      if (translatedTemplate !== messageId) {
+        return translatedTemplate;
+      }
+
+      return viewName.replace('{objectLabelPlural}', objectLabelPlural);
+    }
+
+    if (!isCustom) {
+      const messageId = generateMessageId(viewName);
+      const translatedMessage = this.i18nService.translateMessage({
+        messageId,
+        locale: locale ?? SOURCE_LOCALE,
+      });
+
+      if (translatedMessage !== messageId) {
+        return translatedMessage;
+      }
+    }
+
+    return viewName;
+  }
 
   async findByWorkspaceId(
     workspaceId: string,
@@ -146,187 +479,5 @@ export class ViewService {
     });
 
     return view || null;
-  }
-
-  async create(viewData: Partial<ViewEntity>): Promise<ViewEntity> {
-    if (!isDefined(viewData.workspaceId)) {
-      throw new ViewException(
-        generateViewExceptionMessage(
-          ViewExceptionMessageKey.WORKSPACE_ID_REQUIRED,
-        ),
-        ViewExceptionCode.INVALID_VIEW_DATA,
-        {
-          userFriendlyMessage: generateViewUserFriendlyExceptionMessage(
-            ViewExceptionMessageKey.WORKSPACE_ID_REQUIRED,
-          ),
-        },
-      );
-    }
-
-    if (!isDefined(viewData.objectMetadataId)) {
-      throw new ViewException(
-        generateViewExceptionMessage(
-          ViewExceptionMessageKey.OBJECT_METADATA_ID_REQUIRED,
-        ),
-        ViewExceptionCode.INVALID_VIEW_DATA,
-        {
-          userFriendlyMessage: generateViewUserFriendlyExceptionMessage(
-            ViewExceptionMessageKey.OBJECT_METADATA_ID_REQUIRED,
-          ),
-        },
-      );
-    }
-
-    const view = this.viewRepository.create({
-      ...viewData,
-      isCustom: true,
-    });
-
-    const savedView = await this.viewRepository.save(view);
-
-    await this.flushGraphQLCache(viewData.workspaceId);
-
-    return savedView;
-  }
-
-  async update(
-    id: string,
-    workspaceId: string,
-    updateData: Partial<ViewEntity>,
-    userWorkspaceId?: string,
-  ): Promise<ViewEntity> {
-    const existingView = await this.findById(id, workspaceId);
-
-    if (!isDefined(existingView)) {
-      throw new ViewException(
-        generateViewExceptionMessage(
-          ViewExceptionMessageKey.VIEW_NOT_FOUND,
-          id,
-        ),
-        ViewExceptionCode.VIEW_NOT_FOUND,
-      );
-    }
-
-    // If changing visibility from WORKSPACE to UNLISTED, ensure createdByUserWorkspaceId is set
-    // This prevents the view from disappearing for the user making the change
-    const dataToUpdate = { ...updateData };
-
-    if (
-      isDefined(updateData.visibility) &&
-      updateData.visibility === ViewVisibility.UNLISTED &&
-      existingView.visibility === ViewVisibility.WORKSPACE &&
-      isDefined(userWorkspaceId)
-    ) {
-      // Re-allocate the view to the current user if it has no owner or a different owner
-      dataToUpdate.createdByUserWorkspaceId = userWorkspaceId;
-    }
-
-    const updatedView = await this.viewRepository.save({
-      id,
-      ...dataToUpdate,
-    });
-
-    await this.flushGraphQLCache(workspaceId);
-
-    return { ...existingView, ...updatedView };
-  }
-
-  async delete(id: string, workspaceId: string): Promise<ViewEntity> {
-    const view = await this.findById(id, workspaceId);
-
-    if (!isDefined(view)) {
-      throw new ViewException(
-        generateViewExceptionMessage(
-          ViewExceptionMessageKey.VIEW_NOT_FOUND,
-          id,
-        ),
-        ViewExceptionCode.VIEW_NOT_FOUND,
-      );
-    }
-
-    await this.viewRepository.softDelete(id);
-
-    await this.flushGraphQLCache(workspaceId);
-
-    return view;
-  }
-
-  async destroy(id: string, workspaceId: string): Promise<boolean> {
-    const view = await this.findByIdIncludingDeleted(id, workspaceId);
-
-    if (!isDefined(view)) {
-      throw new ViewException(
-        generateViewExceptionMessage(
-          ViewExceptionMessageKey.VIEW_NOT_FOUND,
-          id,
-        ),
-        ViewExceptionCode.VIEW_NOT_FOUND,
-      );
-    }
-
-    await this.viewRepository.delete(id);
-    await this.flushGraphQLCache(workspaceId);
-
-    return true;
-  }
-
-  processViewNameWithTemplate(
-    viewName: string,
-    isCustom: boolean,
-    objectLabelPlural?: string,
-    locale?: keyof typeof APP_LOCALES,
-  ): string {
-    if (viewName.includes('{objectLabelPlural}') && objectLabelPlural) {
-      const messageId = generateMessageId(viewName);
-      const translatedTemplate = this.i18nService.translateMessage({
-        messageId,
-        values: {
-          objectLabelPlural,
-        },
-        locale: locale ?? SOURCE_LOCALE,
-      });
-
-      if (translatedTemplate !== messageId) {
-        return translatedTemplate;
-      }
-
-      return viewName.replace('{objectLabelPlural}', objectLabelPlural);
-    }
-
-    if (!isCustom) {
-      const messageId = generateMessageId(viewName);
-      const translatedMessage = this.i18nService.translateMessage({
-        messageId,
-        locale: locale ?? SOURCE_LOCALE,
-      });
-
-      if (translatedMessage !== messageId) {
-        return translatedMessage;
-      }
-    }
-
-    return viewName;
-  }
-
-  canUserUpdateView(
-    view: ViewEntity,
-    userWorkspaceId: string | undefined,
-    userHasViewsPermission: boolean,
-  ): boolean {
-    if (userHasViewsPermission) {
-      return true;
-    }
-
-    return (
-      view.visibility === ViewVisibility.UNLISTED &&
-      view.createdByUserWorkspaceId === userWorkspaceId
-    );
-  }
-
-  async flushGraphQLCache(workspaceId: string): Promise<void> {
-    await this.workspaceCacheStorageService.flushGraphQLOperation({
-      operationName: FIND_ALL_CORE_VIEWS_GRAPHQL_OPERATION,
-      workspaceId,
-    });
   }
 }
