+++ packages/twenty-server/src/engine/metadata-modules/view/view.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { ApplicationModule } from 'src/engine/core-modules/application/application.module';
-import { FeatureFlagModule } from 'src/engine/core-modules/feature-flag/feature-flag.module';
 import { I18nModule } from 'src/engine/core-modules/i18n/i18n.module';
 import { WorkspaceManyOrAllFlatEntityMapsCacheModule } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.module';
 import { FlatViewModule } from 'src/engine/metadata-modules/flat-view/flat-view.module';
@@ -17,7 +16,6 @@ import { ViewSortModule } from 'src/engine/metadata-modules/view-sort/view-sort.
 import { ViewController } from 'src/engine/metadata-modules/view/controllers/view.controller';
 import { ViewEntity } from 'src/engine/metadata-modules/view/entities/view.entity';
 import { ViewResolver } from 'src/engine/metadata-modules/view/resolvers/view.resolver';
-import { ViewV2Service } from 'src/engine/metadata-modules/view/services/view-v2.service';
 import { ViewService } from 'src/engine/metadata-modules/view/services/view.service';
 import { WorkspaceCacheStorageModule } from 'src/engine/workspace-cache-storage/workspace-cache-storage.module';
 import { WorkspaceMigrationV2Module } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-v2.module';
@@ -33,7 +31,6 @@ import { WorkspaceMigrationV2Module } from 'src/engine/workspace-manager/workspa
     ViewSortModule,
     I18nModule,
     ApplicationModule,
-    FeatureFlagModule,
     PermissionsModule,
     UserRoleModule,
     WorkspaceCacheStorageModule,
@@ -42,7 +39,7 @@ import { WorkspaceMigrationV2Module } from 'src/engine/workspace-manager/workspa
     WorkspaceManyOrAllFlatEntityMapsCacheModule,
   ],
   controllers: [ViewController],
-  providers: [ViewService, ViewResolver, ViewV2Service],
-  exports: [ViewService, ViewV2Service, TypeOrmModule.forFeature([ViewEntity])],
+  providers: [ViewService, ViewResolver],
+  exports: [ViewService, TypeOrmModule.forFeature([ViewEntity])],
 })
 export class ViewModule {}
