+++ packages/twenty-server/src/engine/workspace-manager/workspace-sync-metadata/constants/default-feature-flags.ts
@@ -1,5 +1,3 @@
 import { FeatureFlagKey } from 'src/engine/core-modules/feature-flag/enums/feature-flag-key.enum';
 
-export const DEFAULT_FEATURE_FLAGS = [
-  FeatureFlagKey.IS_WORKSPACE_MIGRATION_V2_ENABLED,
-] as const satisfies FeatureFlagKey[];
+export const DEFAULT_FEATURE_FLAGS = [] as const satisfies FeatureFlagKey[];
